% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertbar.R
\name{likertbar}
\alias{likertbar}
\title{Plot a percentage stacked bar chart for likert data}
\usage{
likertbar(
  data,
  questions = everything(),
  long_form = FALSE,
  question_col = names(data)[1],
  response_col = names(data)[2],
  count_col = names(data)[3],
  diverging = TRUE,
  colors = NULL,
  palette_name = NULL,
  horizontal = TRUE,
  reverse = FALSE,
  include_na = FALSE,
  na_label = "No Response",
  na_color = "gray",
  order_by = NULL,
  extreme = 1,
  xlab = if (horizontal) "Percent" else "",
  ylab = if (horizontal) "" else "Percent",
  title = ""
)
}
\arguments{
\item{data}{A data.frame object containing information of the likert data.
The data frame can have one of two forms:
(1) Wide form: columns represent questions, rows represent samples, and
each element is a response; or
(2) Long form: a data frame that contains at least the following three
columns:
- \code{Question}: The questions.
- \code{Response}: The responses.
- \code{Count}: The count of each response for each question.
Additional columns in the data frame are allowed and will be ignored by the
function.}

\item{questions}{A character vector of names of columns (questions) that will
be included in the bar chart. Only applicable if data is in wide form.
(default is all columns)}

\item{long_form}{Logical. Indicates if the input data frame is in long form.
If \code{TRUE}, the function assumes the data is in long form; if \code{FALSE}
(default), the function assumes the data is in wide form.}

\item{question_col}{A string specifying the name of the column that contains
the questions. The first column will be used if not provided. Only
applicable if data is in long form.}

\item{response_col}{A string specifying the name of the column that contains
the responses. The second column will be used if not provided. Only
applicable if data is in long form.}

\item{count_col}{A string specifying the name of the column that contains the
count of each response for each question. The third column will be used if
not provided. Only applicable if data is in long form.}

\item{diverging}{Logical. Indicates if the responses are diverging. If \code{TRUE}
(default), the function assumes the responses are diverging; if \code{FALSE},
the function assumes the responses are sequential.}

\item{colors}{A character vector of colors to be used for the different
response types. The number of colors provided must match the number of
unique responses in the data.}

\item{palette_name}{A string specifying the color palette to be used with
\code{RColorBrewer} to generate colors for responses. Ignored if \code{colors} is
provided. If neither \code{colors} nor \code{palette_name} is provided, the default
palettes will be used ("RdBu" if diverging and "Blues" if not).
See \code{RColorBrewer::display.brewer.all()} for available palette names.}

\item{horizontal}{Logical. If \code{TRUE} (default), the bars will be horizontal;
if \code{FALSE}, the bars will be vertical.}

\item{reverse}{Logical. Indicates the sequence of stacked responses based on
response levels. If \code{FALSE} (default), the response with the lower level
appears on the left (if horizontal) or at the bottom (if vertical); if
\code{TRUE}, the order is reversed, with the response with the larger level
appearing on the left or bottom.}

\item{include_na}{Logical. Indicates whether \code{NA} values should be included
in the graph. If \code{TRUE}, \code{NA} will be plotted as a response; If \code{FALSE}
(default), \code{NA} will be dropped.}

\item{na_label}{A string specifying the label to use for \code{NA} values in the
plot. Only applicable if \code{include_na} is \code{TRUE}. (default is "No Response")}

\item{na_color}{A string specifying the color to use for \code{NA} values in the
plot. Only applicable if \code{include_na} is \code{TRUE}. (default is "gray")}

\item{order_by}{A string specifying the criterion used to sort the questions.
Must be \code{"positive"}, \code{"negative"}, \code{"extreme positive"}, \code{"extreme negative"},
or \code{"missing"}. It assumes that the response level is from disagree to agree
(or from negative to positive)
\itemize{
\item \code{"positive"}: Sorts questions by the total count of responses at the
agree levels.
\item \code{"negative"}: Sorts questions by the total count of responses at the
disagree levels.
\item \code{"extreme positive"}: Sorts questions by the count of responses at the
highest levels.
\item \code{"extreme negative"}: Sorts questions by the count of responses at the
lowest levels.
\item \code{"missing"}: Sorts questions by the count of \code{NA} (missing) responses.
If not provided, the questions are plotted in the order they appear in the
data frame.
}}

\item{extreme}{A positive integer indicating the number of extreme levels to
consider if \code{"extreme positive"} or \code{"extreme negative"} is chosen for
\code{order_by}. The default value is 1, meaning only the most extreme level
(e.g., "Strongly Agree" or "Strongly Disagree") will be considered. Higher
values include additional levels toward the center of the response scale.}

\item{xlab, }{ylab, title Strings specifying the labels for the plot:
\itemize{
\item \code{xlab}: The label for the x-axis.
\item \code{ylab}: The label for the y-axis.
\item \code{title}: The main title of the plot.
}}
}
\value{
No return value
}
\description{
This function generates a percentage stacked bar chart for the input likert
data using ggplot2::geom_bar.
}
\examples{
library(tidyverse)
library(likert)
# Load the PISA dataset, which contains Likert-style responses
data("pisaitems")

# Plot a percentage stacked bar chart using the original dataset in wide form
likertbar(pisaitems, questions = c("ST24Q01", "ST24Q02", "ST24Q03"),
          colors = c("#768CE6", "#B3BAE6", "#E8ADB7", "#E16A85"),
          include_na = TRUE, order_by = "extreme positive",
          xlab = "Percentage", ylab = "Questions")

# Transform the data into a long form, with columns "Question", "Response",
# and "Count"
pisaitems_long <- pisaitems |>
  pivot_longer(cols = ST24Q01: ST24Q03, names_to = "Question", values_to = "Response") |>
  group_by(Question, Response) |>
  summarise(Count = n(), .groups = "drop")

# Plot a vertical percentage stacked bar chart using the dataset in long form
likertbar(pisaitems_long, long_form = TRUE, palette_name = "PiYG",
          horizontal = FALSE, order_by = "negative", xlab = "Questions")
}
